+++ __tests__/e2e/dynamic-routes/[id].paths.ts
@@ -1,0 +1,8 @@
+export default {
+  async paths() {
+    return [
+      { params: { id: 'foo' }, content: `# Foo` },
+      { params: { id: 'bar' }, content: `# Bar` }
+    ]
+  }
+}