+++ src/client/app/router.ts
@@ -49,7 +49,7 @@ export function createRouter(
   async function go(href: string = inBrowser ? location.href : '/') {
     await router.onBeforeRouteChange?.(href)
     const url = new URL(href, fakeHost)
-    if (siteDataRef.value.cleanUrls === 'disabled') {
+    if (!siteDataRef.value.cleanUrls) {
       // ensure correct deep link so page refresh lands on correct files.
       // if cleanUrls is enabled, the server should handle this
       if (!url.pathname.endsWith('/') && !url.pathname.endsWith('.html')) {
@@ -57,7 +57,7 @@ export function createRouter(
         href = url.pathname + url.search + url.hash
       }
     }
-    if (inBrowser) {
+    if (inBrowser && href !== location.href) {
       // save scroll position before changing url
       history.replaceState({ scrollPosition: window.scrollY }, document.title)
       history.pushState(null, '', href)
@@ -89,6 +89,18 @@ export function createRouter(
 
         if (inBrowser) {
           nextTick(() => {
+            let actualPathname =
+              siteDataRef.value.base +
+              __pageData.relativePath.replace(/(?:(^|\/)index)?\.md$/, '$1')
+            if (!siteDataRef.value.cleanUrls && !actualPathname.endsWith('/')) {
+              actualPathname += '.html'
+            }
+            if (actualPathname !== targetLoc.pathname) {
+              targetLoc.pathname = actualPathname
+              href = actualPathname + targetLoc.search + targetLoc.hash
+              history.replaceState(null, '', href)
+            }
+
             if (targetLoc.hash && !scrollPosition) {
               let target: HTMLElement | null = null
               try {
@@ -141,9 +153,21 @@ export function createRouter(
         const button = (e.target as Element).closest('button')
         if (button) return
 
-        const link = (e.target as Element).closest('a')
-        if (link && !link.closest('.vp-raw') && !link.download) {
-          const { href, origin, pathname, hash, search, target } = link
+        const link = (e.target as Element | SVGElement).closest<
+          HTMLAnchorElement | SVGAElement
+        >('a')
+        if (
+          link &&
+          !link.closest('.vp-raw') &&
+          (link instanceof SVGElement || !link.download)
+        ) {
+          const { target } = link
+          const { href, origin, pathname, hash, search } = new URL(
+            link.href instanceof SVGAnimatedString
+              ? link.href.animVal
+              : link.href,
+            link.baseURI
+          )
           const currentUrl = window.location
           const extMatch = pathname.match(/\.\w+$/)
           // only intercept inbound links
@@ -205,8 +229,8 @@ export function useRoute(): Route {
   return useRouter().route
 }
 
-function scrollTo(el: HTMLElement, hash: string, smooth = false) {
-  let target: HTMLElement | null = null
+function scrollTo(el: HTMLElement | SVGElement, hash: string, smooth = false) {
+  let target: HTMLElement | SVGElement | null = null
 
   try {
     target = el.classList.contains('header-anchor')