+++ src/client/theme-default/components/VPFeature.vue
@@ -7,7 +7,7 @@ import VPIconArrowRight from './icons/VPIconArrowRight.vue'
 defineProps<{
   icon?: DefaultTheme.FeatureIcon
   title: string
-  details: string
+  details?: string
   link?: string
   linkText?: string
 }>()
@@ -24,8 +24,8 @@ defineProps<{
         :width="icon.width"
       />
       <div v-else-if="icon" class="icon">{{ icon }}</div>
-      <h2 class="title">{{ title }}</h2>
-      <p class="details">{{ details }}</p>
+      <h2 class="title" v-html="title"></h2>
+      <p v-if="details" class="details" v-html="details"></p>
 
       <div v-if="linkText" class="link-text">
         <p class="link-text-value">