+++ src/client/theme-default/components/VPLink.vue
@@ -5,21 +5,25 @@ import VPIconExternalLink from './icons/VPIconExternalLink.vue'
 import { EXTERNAL_URL_RE } from '../../shared.js'
 
 const props = defineProps<{
+  tag?: string
   href?: string
   noIcon?: boolean
+  target?: string
+  rel?: string
 }>()
 
+const tag = computed(() => props.tag ?? props.href ? 'a' : 'span')
 const isExternal = computed(() => props.href && EXTERNAL_URL_RE.test(props.href))
 </script>
 
 <template>
   <component
-    :is="href ? 'a' : 'span'"
+    :is="tag"
     class="VPLink"
     :class="{ link: href }"
     :href="href ? normalizeLink(href) : undefined"
-    :target="isExternal ? '_blank' : undefined"
-    :rel="isExternal ? 'noreferrer' : undefined"
+    :target="target || (isExternal ? '_blank' : undefined)"
+    :rel="rel || (isExternal ? 'noreferrer' : undefined)"
   >
     <slot />
     <VPIconExternalLink v-if="isExternal && !noIcon" class="icon" />
@@ -35,5 +39,6 @@ const isExternal = computed(() => props.href && EXTERNAL_URL_RE.test(props.href)
   height: 11px;
   fill: var(--vp-c-text-3);
   transition: fill 0.25s;
+  flex-shrink: 0;
 }
 </style>