+++ src/client/theme-default/components/VPNavBarSearch.vue
@@ -29,10 +29,26 @@ const buttonText = computed(
     'Search'
 )
 
+const preconnect = () => {
+  const id = 'VPAlgoliaPreconnect'
+
+  const rIC = window.requestIdleCallback || setTimeout
+  rIC(() => {
+    const preconnect = document.createElement('link')
+    preconnect.id = id
+    preconnect.rel = 'preconnect'
+    preconnect.href = `https://${theme.value.algolia!.appId}-dsn.algolia.net`
+    preconnect.crossOrigin = ''
+    document.head.appendChild(preconnect)
+  })
+}
+
 onMounted(() => {
   if (!theme.value.algolia) {
     return
   }
+
+  preconnect()
 
   // meta key detect (same logic as in @docsearch/js)
   metaKey.value = /(Mac|iPhone|iPod|iPad)/i.test(navigator.platform)