+++ src/client/theme-default/composables/prev-next.ts
@@ -15,12 +15,39 @@ export function usePrevNext() {
     })
 
     return {
-      prev: frontmatter.value.prev
-        ? { ...candidates[index - 1], text: frontmatter.value.prev }
-        : candidates[index - 1],
-      next: frontmatter.value.next
-        ? { ...candidates[index + 1], text: frontmatter.value.next }
-        : candidates[index + 1]
+      prev:
+        frontmatter.value.prev === false
+          ? undefined
+          : {
+              text:
+                (typeof frontmatter.value.prev === 'string'
+                  ? frontmatter.value.prev
+                  : typeof frontmatter.value.prev === 'object'
+                  ? frontmatter.value.prev.text
+                  : undefined) ?? candidates[index - 1]?.text,
+              link:
+                (typeof frontmatter.value.prev === 'object'
+                  ? frontmatter.value.prev.link
+                  : undefined) ?? candidates[index - 1]?.link
+            },
+      next:
+        frontmatter.value.next === false
+          ? undefined
+          : {
+              text:
+                (typeof frontmatter.value.next === 'string'
+                  ? frontmatter.value.next
+                  : typeof frontmatter.value.next === 'object'
+                  ? frontmatter.value.next.text
+                  : undefined) ?? candidates[index + 1]?.text,
+              link:
+                (typeof frontmatter.value.next === 'object'
+                  ? frontmatter.value.next.link
+                  : undefined) ?? candidates[index + 1]?.link
+            }
+    } as {
+      prev?: { text?: string; link?: string }
+      next?: { text?: string; link?: string }
     }
   })
 }