+++ src/client/theme-default/composables/sidebar.ts
@@ -1,15 +1,32 @@
 import {
+  type ComputedRef,
+  type Ref,
   computed,
   onMounted,
   onUnmounted,
-  type Ref,
   ref,
   watchEffect
 } from 'vue'
-import { useRoute } from 'vitepress'
 import { useMediaQuery } from '@vueuse/core'
+import { useRoute } from 'vitepress'
+import type { DefaultTheme } from 'vitepress/theme'
+import { isActive } from '../support/utils.js'
+import {
+  hasActiveLink as containsActiveLink,
+  getSidebar,
+  getSidebarGroups
+} from '../support/sidebar.js'
 import { useData } from './data.js'
-import { getSidebar } from '../support/sidebar.js'
+
+export interface SidebarControl {
+  collapsed: Ref<boolean>
+  collapsible: ComputedRef<boolean>
+  isLink: ComputedRef<boolean>
+  isActiveLink: ComputedRef<boolean>
+  hasActiveLink: ComputedRef<boolean>
+  hasChildren: ComputedRef<boolean>
+  toggle(): void
+}
 
 export function useSidebar() {
   const route = useRoute()
@@ -33,12 +50,17 @@ export function useSidebar() {
   })
 
   const hasAside = computed(() => {
-    return (
-      frontmatter.value.layout !== 'home' && frontmatter.value.aside !== false
-    )
+    if (frontmatter.value.layout === 'home') return false
+    if (frontmatter.value.aside != null) return !!frontmatter.value.aside
+    if (theme.value.aside === false) return false
+    return true
   })
 
   const isSidebarEnabled = computed(() => hasSidebar.value && is960.value)
+
+  const sidebarGroups = computed(() => {
+    return hasSidebar.value ? getSidebarGroups(sidebar.value) : []
+  })
 
   function open() {
     isOpen.value = true
@@ -55,6 +77,7 @@ export function useSidebar() {
   return {
     isOpen,
     sidebar,
+    sidebarGroups,
     hasSidebar,
     hasAside,
     isSidebarEnabled,
@@ -93,5 +116,63 @@ export function useCloseSidebarOnEscape(
       close()
       triggerElement?.focus()
     }
+  }
+}
+
+export function useSidebarControl(
+  item: ComputedRef<DefaultTheme.SidebarItem>
+): SidebarControl {
+  const { page } = useData()
+
+  const collapsed = ref(false)
+
+  const collapsible = computed(() => {
+    return item.value.collapsed != null
+  })
+
+  const isLink = computed(() => {
+    return !!item.value.link
+  })
+
+  const isActiveLink = computed(() => {
+    return isActive(page.value.relativePath, item.value.link)
+  })
+
+  const hasActiveLink = computed(() => {
+    if (isActiveLink.value) {
+      return true
+    }
+
+    return item.value.items
+      ? containsActiveLink(page.value.relativePath, item.value.items)
+      : false
+  })
+
+  const hasChildren = computed(() => {
+    return !!(item.value.items && item.value.items.length)
+  })
+
+  watchEffect(() => {
+    collapsed.value = !!(collapsible.value && item.value.collapsed)
+  })
+
+  watchEffect(() => {
+    ;(isActiveLink.value || hasActiveLink.value) && (collapsed.value = false)
+  })
+
+  function toggle() {
+    if (collapsible.value) {
+      collapsed.value = !collapsed.value
+    }
+  }
+
+  return {
+    collapsed,
+    collapsible,
+    isLink,
+    isActiveLink,
+    hasActiveLink,
+    hasChildren,
+    toggle
   }
 }